+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -21,8 +21,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static kitchenpos.application.MenuFixture.*;
-import static kitchenpos.application.ProductFixture.createProduct;
+import static kitchenpos.application.fixture.MenuFixture.*;
+import static kitchenpos.application.fixture.ProductFixture.createProduct;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
