+++ src/test/java/kitchenpos/application/fixture/MenuGroupFixture.java
@@ -1,15 +1,15 @@
-package kitchenpos.application;
+package kitchenpos.application.fixture;
 
 import kitchenpos.domain.MenuGroup;
 
 public class MenuGroupFixture {
-    static MenuGroup createMenuGroupRequest(String name) {
+    public static MenuGroup createMenuGroupRequest(String name) {
         MenuGroup menuGroup = new MenuGroup();
         menuGroup.setName(name);
         return menuGroup;
     }
 
-    static MenuGroup createMenuGroup(Long id, String name) {
+    public static MenuGroup createMenuGroup(Long id, String name) {
         MenuGroup menuGroup = new MenuGroup();
         menuGroup.setId(id);
         menuGroup.setName(name);
