+++ src/test/java/kitchenpos/application/fixture/OrderFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.fixture;
 
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
@@ -7,7 +7,7 @@ import kitchenpos.domain.OrderStatus;
 import java.util.List;
 
 public class OrderFixture {
-    static Order createOrder(Long id, OrderStatus status, Long orderTableId) {
+    public static Order createOrder(Long id, OrderStatus status, Long orderTableId) {
         Order order = new Order();
         order.setOrderStatus(status.name());
         order.setId(id);
@@ -15,14 +15,14 @@ public class OrderFixture {
         return order;
     }
 
-    static Order createOrderRequest(List<OrderLineItem> orderLineItems, Long orderTableId) {
+    public static Order createOrderRequest(List<OrderLineItem> orderLineItems, Long orderTableId) {
         Order order = new Order();
         order.setOrderLineItems(orderLineItems);
         order.setOrderTableId(orderTableId);
         return order;
     }
 
-    static Order updateOrderRequest(OrderStatus status) {
+    public static Order updateOrderRequest(OrderStatus status) {
         Order order = new Order();
         order.setOrderStatus(status.name());
         return order;
