+++ src/test/java/kitchenpos/application/fixture/ProductFixture.java
@@ -1,18 +1,18 @@
-package kitchenpos.application;
+package kitchenpos.application.fixture;
 
 import kitchenpos.domain.Product;
 
 import java.math.BigDecimal;
 
 public class ProductFixture {
-    static Product createProductRequest(String name, BigDecimal price) {
+    public static Product createProductRequest(String name, BigDecimal price) {
         Product product = new Product();
         product.setName(name);
         product.setPrice(price);
         return product;
     }
 
-    static Product createProduct(Long id, String name, BigDecimal price) {
+    public static Product createProduct(Long id, String name, BigDecimal price) {
         Product product = new Product();
         product.setId(id);
         product.setName(name);
