+++ src/test/java/kitchenpos/application/fixture/TableGroupFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.fixture;
 
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class TableGroupFixture {
-    static TableGroup createTableGroupRequest(List<Long> orderTableIds) {
+    public static TableGroup createTableGroupRequest(List<Long> orderTableIds) {
         TableGroup tableGroup = new TableGroup();
         List<OrderTable> orderTables = orderTableIds.stream()
                 .map(it -> new OrderTable() {{
@@ -18,7 +18,7 @@ public class TableGroupFixture {
         return tableGroup;
     }
 
-    static TableGroup createTableGroup(List<Long> orderTableIds) {
+    public static TableGroup createTableGroup(List<Long> orderTableIds) {
         TableGroup tableGroup = new TableGroup();
         List<OrderTable> orderTables = orderTableIds.stream()
                 .map(it -> new OrderTable() {{
